Refactor Navigation component into TopMar menu.
